git-svn-id: file:///home/xapian-svn/xapian/branches/gsoc2009-kosei@14940 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/chert/chert_types.h
@@ -42,7 +42,7 @@ typedef unsigned int chert_revision_number_t;
  *
  *  Again, this must be of suitable minimum size.
  */
-typedef unsigned int chert_tablesize_t;
+typedef unsigned long long chert_tablesize_t;
 
 /** An integer type for storing the length of a document - ie, the sum of the
  *  wdfs of the terms in the document.
